backend_drf/requirements.txt
amqp==5.3.1
asgiref==3.9.0
async-timeout==5.0.1
billiard==4.2.1
celery==5.5.3
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cron-descriptor==1.4.5
cryptography==45.0.5
defusedxml==0.7.1
Django==5.2.4
django-autoslug==1.9.9
django-celery-beat==2.8.1
django-countries==7.6.1
django-decouple==2.1
django-filter==25.1
django-phonenumber-field==8.1.0
django-redis==6.0.0
django-timezone-field==7.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
djoser==2.3.1
drf-yasg==1.21.10
flower==2.0.1
humanize==4.12.3
idna==3.10
inflection==0.5.1
kombu==5.5.4
oauthlib==3.3.1
packaging==25.0
phonenumbers==9.0.8
pillow==11.3.0
prometheus_client==0.22.1
prompt_toolkit==3.0.51
pycparser==2.22
PyJWT==2.9.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.2
redis==6.2.0
requests==2.32.4
requests-oauthlib==2.0.0
six==1.17.0
social-auth-app-django==5.5.1
social-auth-core==4.7.0
sqlparse==0.5.3
tornado==6.5.1
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
